+++ NEWS.md
@@ -1,6 +1,7 @@
 # tipr 0.4.1
 
 * Add additional functions that specify `*_with_continuous()` (long form of, the function names, the default unmeasured confounder is Normally distributed)
+* Change `tip_lm()` to `tip_coef()`.
 
 # tipr 0.4.0